% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_scale.R
\name{min_max_scale}
\alias{min_max_scale}
\title{Scaling by Min and Max}
\usage{
min_max_scale(x)
}
\arguments{
\item{x}{A numic vector}
}
\value{
the same vector of numbers, but scaled by the minimum and maximum in the column
}
\description{
Scaling by Min and Max
}
\examples{
min_max_scale(c(1, 2, 3))
}
