% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_academic.R
\name{theme_academic}
\alias{theme_academic}
\title{Academic Theme}
\usage{
theme_academic(base_size = 12)
}
\arguments{
\item{base_size}{A positive integer}
}
\value{
A plot with the theme added when passed through ggplot2 (centralized plot title)
}
\description{
Academic Theme
}
\examples{
library(ggplot2)
p <- ggplot(data = mtcars) + geom_point(aes(x = mpg, y = hp)) + labs(title = "plot_title")
p + theme_academic()
}
