% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{A toy dataset for demonstrating the functions in toypackage}
\format{
\subsection{\code{toy_data}}{

A data frame with 5 rows and 4 columns:
\describe{
\item{subject_id}{An integer denoting the subject id}
\item{reaction_time}{Reaction time in seconds}
\item{test_score}{Test score of each subject}
\item{recalled_words}{Words recalled by each subject}
}
}
}
\usage{
toy_data
}
\description{
A toy dataset for demonstrating the functions in toypackage
}
\keyword{datasets}
