% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a String}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vector with one element}

\item{split}{A character to split on}
}
\value{
A character vector
}
\description{
Split a String
}
\examples{
strsplit1("alpha,bro", split = ",")
}
